Create a new animation component page-scroll and import in homepage
