+++ src/app/page.tsx
@@ -1,59 +1,14 @@
 "use client";
-import { useState, useEffect } from "react";
-import { useRouter } from "next/navigation";
-import { motion } from "framer-motion";
 import { TypeAnimation } from "react-type-animation";
 
 import FullscreenMenu from "@/components/navigation/fullscreen-menu";
 import Image from "next/image";
 import { ArrowDown } from "lucide-react";
+import PageScroll from "@/components/animations/page-scroll";
 
 export default function HomePage() {
-  // Smooth scrolling to AboutPage
-  const router = useRouter();
-  const [scrolling, setScrolling] = useState(false);
-  const [touchStartY, setTouchStartY] = useState(0);
-
-  useEffect(() => {
-    router.prefetch("/about"); // Preload AboutPage
-  }, [router]);
-
-  // Handle scroll on desktop-device
-  const handleScroll = () => {
-    if (!scrolling) {
-      setScrolling(true);
-      setTimeout(() => {
-        router.push("/about");
-      }, 500);
-    }
-  };
-
-  // Handle touch on mobile-device
-  const handleTouchStart = (e: React.TouchEvent) => {
-    setTouchStartY(e.touches[0].clientY);
-  };
-
-  const handleTouchMove = (e: React.TouchEvent) => {
-    const touchEndY = e.touches[0].clientY;
-    const deltaY = touchStartY - touchEndY;
-
-    if (deltaY > 50) {
-      // Swipe up
-      handleScroll();
-    }
-  };
-
   return (
-    <motion.div
-      className="relative w-full h-screen"
-      animate={
-        scrolling ? { scale: 0.5, opacity: 0 } : { scale: 1, opacity: 1 }
-      }
-      transition={{ duration: 0.7 }}
-      onWheel={handleScroll} // Desktop
-      onTouchStart={handleTouchStart} // Mobile
-      onTouchMove={handleTouchMove} // Mobile
-    >
+    <PageScroll>
       {/* Hero section */}
       <header className="flex flex-col items-center justify-center relative w-screen h-screen">
         <div className="z-1 absolute top-0 right-0 p-5">
@@ -96,6 +51,6 @@ export default function HomePage() {
           </div>
         </section>
       </header>
-    </motion.div>
+    </PageScroll>
   );
 }
